Revert "update: スレッド関連をカスタムhook化"
This reverts commit 25c2f5b3cb702f11d0dc0e37f5aef353e163219d.
